fix(render): round an intensity to float32 before comparing it
The review found the "only what changed is uploaded" promise defeated
by the one value the demo actually dims with. A Float32Array reads back
the rounded double, so a caller's 0.2 never equals the
0.20000000298023224 that came out of it: every dimmed edge re-marked
itself on every hover, the merged span grew to the whole buffer, and
the ranges cost more than they saved. The tests missed it because 1,
0.25 and 0.5 are all exact in float32.

Rounding at the boundary fixes it, and the new test uses 0.2 and reads
the attribute VERSION rather than its range list: no backend has run in
a test, so a flush that finds nothing dirty leaves the previous range
in place and the list cannot tell the two cases apart.

Two more from the same review, both about a number a screenshot
testifies to. The readout counted far-end REGISTRATIONS where the
committed frame drew eight of nine names, because a neighbour that
already has a title of its own is gated out of the far-end tier; it
counts what is on screen now. And the hover frame's caption still
described the node the frame was scouted against, claiming nineteen
edges where the readout in the picture says nine, so both frames are
retaken with captions that match them.

The D3 frame-set docblock had been inserted above D2's key, leaving D2
undocumented and D3 carrying a paragraph about D2's gates. Moved.

+++ packages/render/src/scene-edges.ts
@@ -364,7 +364,17 @@ class EdgeGroup implements GpuResource {
   /** Writes one intensity per edge. See {@link SceneEdges.setEdgeIntensity}. */
   setIntensity(intensityOf: (edgeId: string) => number): void {
     for (const range of this.#ranges) {
-      const value = requireIntensity(intensityOf(range.id), `intensity of edge "${range.id}"`);
+      // ROUNDED TO FLOAT32 BEFORE ANYTHING COMPARES IT, and this line is the
+      // whole of the "only what changed is uploaded" promise. The array stores
+      // float32 and reads back the rounded double, so a caller's 0.2 never
+      // equals the 0.20000000298023224 that came out of it: without the round,
+      // every value that is not exactly representable re-marks every one of its
+      // vertices on every call, the merged span grows to the whole buffer, and
+      // the ranges cost more than they save. A review caught it against the one
+      // number the campaign demo actually dims with.
+      const value = Math.fround(
+        requireIntensity(intensityOf(range.id), `intensity of edge "${range.id}"`),
+      );
       const end = range.vertexStart + range.vertexCount;
       // Compared before written, so a call that changes nothing uploads
       // nothing. That is the common case by construction: a hover moving from
